Streamlit_Object_Detection/requirements.txt
streamlit
ultralytics
opencv-python
pillow
numpy
requests
transformers
torch